requirements.txt
Django==1.9.11
django-betterforms==1.1.4
django-braces==1.9.0
django-mssql==1.7
django-pure-pagination==0.3.0
django-pyodbc==0.3.0
django-pyodbc-azure==1.9.3.0
django-widget-tweaks==1.4.1
djangorestframework==3.4.5
Pillow==3.4.2
pyodbc==3.0.10
qrcode==5.3
six==1.10.0
sql-server.pyodbc==1.0
virtualenv==15.0.3
